# docs/openapi-pre-decomposition.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stitchd Feature Flag API",
    "description": "REST API for managing feature flags, segments, and evaluation.",
    "contact": {
      "name": "Stitchd"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/environments/{env_id}/experiments": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments` — List experiments in an environment.",
        "description": "# Errors\nReturns 500 on database error.",
        "operationId": "list_experiments",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of experiments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "`POST /v1/environments/{env_id}/experiments` — Create a new experiment.",
        "description": "# Errors\nReturns 422 if validation fails; 500 on database error.",
        "operationId": "create_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Experiment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments/{id}` — Get a single experiment.",
        "description": "# Errors\nReturns 404 if not found; 500 on database error.",
        "operationId": "get_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "`DELETE /v1/environments/{env_id}/experiments/{id}` — Soft-delete an experiment.",
        "description": "# Errors\nReturns 404 if not found; 409 if experiment is currently running; 500 on database error.",
        "operationId": "delete_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Experiment deleted"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict — experiment is running"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "patch": {
        "tags": [
          "experiments"
        ],
        "summary": "`PATCH /v1/environments/{env_id}/experiments/{id}` — Update an experiment.",
        "description": "# Errors\nReturns 404 if not found; 409 on version conflict or invalid state; 500 on database error.",
        "operationId": "update_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}/iterations": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments/{id}/iterations` — List iterations for an experiment.",
        "description": "# Errors\nReturns 404 if not found; 500 on database error.",
        "operationId": "list_iterations",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of iterations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentIterationResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}/transitions": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "`POST /v1/environments/{env_id}/experiments/{id}/transitions` — Transition experiment lifecycle.",
        "description": "# Errors\nReturns 404 if not found; 409 on invalid transition; 500 on database error.",
        "operationId": "transition_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment transitioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Invalid transition"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateExperimentRequest": {
        "type": "object",
        "description": "Request body for creating a new experiment.",
        "required": [
          "flag_rule_id",
          "name",
          "metric_keys"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description of what the experiment tests."
          },
          "flag_rule_id": {
            "$ref": "#/components/schemas/RuleId",
            "description": "The flag rule this experiment is bound to."
          },
          "hypothesis": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional hypothesis statement for the experiment."
          },
          "metric_keys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Pre-registered event definition keys to use as metrics (at least one required)."
          },
          "min_sample_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional informational minimum sample size guardrail."
          },
          "name": {
            "type": "string",
            "description": "Human-readable name for the experiment."
          },
          "scheduled_end_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional ISO-8601 scheduled end time (informational only)."
          },
          "scheduled_start_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional ISO-8601 scheduled start time (informational only)."
          },
          "traffic_allocation": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Percentage of rule-matched contexts to enrol (0.1–100.0); defaults to 100.0."
          }
        }
      },
...
      "EnvironmentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ExperimentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ExperimentResponse": {
        "type": "object",
        "description": "Response body for a single experiment.",
        "required": [
          "id",
          "env_id",
          "flag_rule_id",
          "name",
          "status",
          "metric_keys",
          "traffic_allocation",
          "version",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description of the experiment."
          },
          "env_id": {
            "$ref": "#/components/schemas/EnvironmentId",
            "description": "The environment this experiment belongs to."
          },
          "flag_rule_id": {
            "$ref": "#/components/schemas/RuleId",
            "description": "The flag rule this experiment is bound to."
          },
          "hypothesis": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional hypothesis statement."
          },
          "id": {
            "$ref": "#/components/schemas/ExperimentId",
            "description": "Unique identifier of the experiment."
          },
          "metric_keys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Pre-registered event definition keys used as metrics."
          },
          "min_sample_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional informational minimum sample size guardrail."
          },
          "name": {
            "type": "string",
            "description": "Human-readable name."
          },
          "scheduled_end_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional ISO-8601 scheduled end time."
          },
          "scheduled_start_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional ISO-8601 scheduled start time."
          },
          "status": {
            "type": "string",
            "description": "Current lifecycle status serialized as a `snake_case` string."
          },
          "traffic_allocation": {
            "type": "number",
            "format": "double",
            "description": "Percentage of rule-matched contexts enrolled (0.1–100.0)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 last-updated timestamp."
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "Optimistic-concurrency version counter."
          }
        }
      },
...
      "RuleId": {
        "type": "string",
        "format": "uuid"
      },
...
      "UpdateExperimentRequest": {
        "type": "object",
        "description": "Request body for updating an existing experiment.",
        "required": [
          "version"
        ],
        "properties": {
          "analysis_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Statistical analysis type (`\"frequentist\"` or `\"bayesian\"`).\n\n**Cannot be changed while the experiment is in `running` state** — doing\nso would invalidate already-computed results.  A 409 Conflict is returned\nif this field is supplied for a running experiment."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Updated description of the experiment."
          },
          "hypothesis": {
            "type": [
              "string",
              "null"
            ],
            "description": "Updated hypothesis statement."
          },
          "metric_keys": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Updated list of metric event definition keys (at least one required if provided)."
          },
          "min_sample_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Updated minimum sample size guardrail."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Updated human-readable name for the experiment."
          },
          "scheduled_end_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Updated scheduled end time."
          },
          "scheduled_start_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Updated scheduled start time."
          },
          "traffic_allocation": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Updated traffic allocation percentage (0.1–100.0)."
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "Current version for optimistic locking (required)."
          }
        }
      },
...
    }
  }
}